[Patch] support backing image in ramfs
When an image is opened with directsync/none mode and its backing
file is located in ramfs such as /dev/shm, the drive can not be
opened.

This patch tries to remove O_DIRECT flag of a backing file if it
fails to open.
(cherry picked from commit a9e69b5b60c22a17913ad7dee213831d143b2a04)